+++ openssl/crypto/x509/x509_trs.c
@@ -123,7 +123,7 @@ int X509_check_trust(X509 *x, int id, int flags)
 int X509_TRUST_get_count(void)
 {
 	if(!trtable) return X509_TRUST_COUNT;
-	return sk_X509_TRUST_num(trtable) + X509_TRUST_COUNT;
+	return (int)(sk_X509_TRUST_num(trtable) + X509_TRUST_COUNT);
 }
 
 X509_TRUST * X509_TRUST_get0(int idx)
@@ -143,7 +143,7 @@ int X509_TRUST_get_by_id(int id)
 	if(!trtable) return -1;
 	idx = sk_X509_TRUST_find(trtable, &tmp);
 	if(idx == -1) return -1;
-	return idx + X509_TRUST_COUNT;
+	return (int)(idx + X509_TRUST_COUNT);
 }
 
 int X509_TRUST_set(int *t, int trust)
