+++ openssl/ssl/s23_clnt.c
@@ -150,7 +150,7 @@ int ssl23_connect(SSL *s)
 	int ret= -1;
 	int new_state,state;
 
-	RAND_add(&Time,sizeof(Time),0);
+	RAND_add(&Time,sizeof(Time),(double)0);
 	ERR_clear_error();
 	clear_sys_error();
