+++ openssl/ssl/s2_clnt.c
@@ -150,7 +150,7 @@ int ssl2_connect(SSL *s)
 	void (*cb)(const SSL *ssl,int type,int val)=NULL;
 	int new_state,state;
 
-	RAND_add(&l,sizeof(l),0);
+	RAND_add(&l,sizeof(l),(double)0);
 	ERR_clear_error();
 	clear_sys_error();
