+++ openssl/ssl/s3_clnt.c
@@ -191,7 +191,7 @@ int ssl3_connect(SSL *s)
 	int ret= -1;
 	int new_state,state,skip=0;
 
-	RAND_add(&Time,sizeof(Time),0);
+	RAND_add(&Time,sizeof(Time),(double)0);
 	ERR_clear_error();
 	clear_sys_error();
